+++ Nernel/Source/CLR/InternalCalls/System.DateTime.c
@@ -2,7 +2,7 @@
 #include <CLR/InternalCalls/System.DateTime.h>
 #include <CLR/InternalCalls/Helpers.h>
 
-uint32_t System_DateTime_InternalUtcNow(AppDomain* pAppDomain)
+uint64_t System_DateTime_InternalUtcNow(AppDomain* pAppDomain)
 {
-	return (uint32_t)SystemClock_GetTicks();
+	return (uint64_t)SystemClock_GetTicks();
 }
